+++ src/commonMain/kotlin/CallbackKSLog.kt
@@ -1,7 +1,9 @@
 package dev.inmo.kslog.common
 
+typealias SimpleKSLogCallback = (level: LogLevel, tag: String?, message: Any, throwable: Throwable?) -> Unit
+
 class CallbackKSLog(
-    private val performLogCallback: (level: LogLevel, tag: String?, message: Any, throwable: Throwable?) -> Unit
+    private val performLogCallback: SimpleKSLogCallback
 ) : KSLog {
     override fun performLog(level: LogLevel, tag: String?, message: Any, throwable: Throwable?) = performLogCallback(level, tag, message, throwable)
 }